refactor(slack): simplify integration.definition.ts
