+++ integrations/slack/src/definitions/channels.ts
@@ -1,4 +1,4 @@
-import { IntegrationDefinitionProps, messages as baseMessages } from '@botpress/sdk'
+import { messages as baseMessages } from '@botpress/sdk'
 import { textSchema } from './schemas'
 
 const messages = {
@@ -13,32 +13,32 @@ const convoTags = {
   title: {},
 } as const
 
-export const channels = {
-  channel: {
-    title: 'Channel',
-    messages,
-    message: { tags: { ts: {} } },
-    conversation: {
-      tags: { ...convoTags },
-      creation: { enabled: true, requiredTags: ['id'] },
-    },
+export const channel = {
+  title: 'Channel',
+  messages,
+  message: { tags: { ts: {} } },
+  conversation: {
+    tags: { ...convoTags },
+    creation: { enabled: true, requiredTags: ['id'] },
   },
-  dm: {
-    title: 'Direct Message',
-    messages,
-    message: { tags: { ts: {} } },
-    conversation: {
-      tags: { ...convoTags },
-      creation: { enabled: true, requiredTags: ['id'] },
-    },
+}
+
+export const dm = {
+  title: 'Direct Message',
+  messages,
+  message: { tags: { ts: {} } },
+  conversation: {
+    tags: { ...convoTags },
+    creation: { enabled: true, requiredTags: ['id'] },
   },
-  thread: {
-    title: 'Thread',
-    messages,
-    message: { tags: { ts: {} } },
-    conversation: {
-      tags: { ...convoTags, thread: {} },
-      creation: { enabled: true, requiredTags: ['id'] },
-    },
+}
+
+export const thread = {
+  title: 'Thread',
+  messages,
+  message: { tags: { ts: {} } },
+  conversation: {
+    tags: { ...convoTags, thread: {} },
+    creation: { enabled: true, requiredTags: ['id'] },
   },
-} satisfies IntegrationDefinitionProps['channels']
+}
